MOTECH-2084: Moved all contract related classes to the contract package
